% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbbga_2023}
\alias{cbbga_2023}
\title{College Basketball Games Data Set for 2023 Season}
\format{
A data frame with 6 variables: \code{Date}, \code{Away_Team},
\code{Away_Score}, \code{Home_Team}, \code{Home_Score}, and \code{Score_Difference}.
}
\usage{
cbbga_2023
}
\description{
This data set contains information on college basketball games for the 2023 season,
including the date, away team, away team score, home team, home team score, and
score difference.
}
\keyword{datasets}
