% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_by_team.R
\name{games_by_team}
\alias{games_by_team}
\title{Filter games by a given team}
\usage{
games_by_team(data, team_name)
}
\arguments{
\item{data}{A tibble containing basketball games data.}

\item{team_name}{The name of the team for which games are to be filtered.}
}
\value{
A tibble containing only the games played by the specified team.
}
\description{
This function filters a tibble containing basketball games data to include only
the games played by a specified team.
}
\examples{
# Filter games for the team named "Duke"
duke_games <- games_by_team(cbbga, "Duke")

}
