% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_shiny_app.R
\name{launch_shiny_app}
\alias{launch_shiny_app}
\title{Launches the Shiny app included in the package.}
\usage{
launch_shiny_app()
}
\description{
This function launches the Shiny app that is included in the package.
Users can interact with the Shiny app by entering a team name to view
statistics and plots related to that team.
}
\examples{
\dontrun{
launch_shiny_app()
}
}
