% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_stats.R
\name{team_stats}
\alias{team_stats}
\title{Compute win/loss record and winning percentage for a given team}
\usage{
team_stats(data, team_name)
}
\arguments{
\item{data}{A tibble containing basketball game data, including dates, teams, and scores.}

\item{team_name}{A character string specifying the name of the team for which statistics are to be computed.}
}
\value{
A tibble with the following columns:
\describe{
\item{Team}{The name of the team.}
\item{Wins}{The number of wins for the team.}
\item{Losses}{The number of losses for the team.}
\item{Total_Games}{The total number of games played by the team.}
\item{Winning_Percentage}{The winning percentage of the team, calculated as the ratio of wins to total games.}
}
}
\description{
This function calculates the win/loss record and winning percentage for a specified team based on the provided data.
}
\examples{
# Compute team statistics for "Southern Utah"
team_stats(cbbga_2023, "Southern Utah")

}
