% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_games_histogram.R
\name{total_games_histogram}
\alias{total_games_histogram}
\title{Create Histogram of Games for each team}
\usage{
total_games_histogram(data)
}
\arguments{
\item{data}{A tibble containing basketball game data, including dates, teams, and scores.}
}
\value{
A ggplot object representing the histogram of total games played by each team.
}
\description{
This function generates a histogram showing the distribution of total games played by each team.
}
\examples{
# Generate histogram of total games for the provided data
total_games_histogram(cbbga_2023)

}
